:fire: remove win32 support as it has no pipe

+++ tests/integration_tests/test_command_line_options.py
@@ -499,6 +499,8 @@ def test_pypi_pkg_example(_):
 
 
 def test_add_extension():
+    if sys.version_info[0] == 2:
+        raise SkipTest("jinja2-python-version does not support python 2")
     test_commands = [
         [
             "moban",
@@ -526,4 +528,19 @@ def test_add_extension():
                     content,
                     "{}.{}".format(sys.version_info[0], sys.version_info[1]),
                 )
+            os.unlink("moban.output")
+
+
+def test_stdin_input():
+    if sys.version_info[0] == 2:
+        raise SkipTest("windows test fails with this pipe test 2")
+    test_args = ["moban", "-d", "hello=world"]
+    with patch.object(sys, "stdin", StringIO("{{hello}}")):
+        with patch.object(sys, "argv", test_args):
+            from moban.main import main
+
+            main()
+            with open("moban.output") as f:
+                content = f.read()
+                eq_(content, "world")
             os.unlink("moban.output")
